Fixing stuff

+++ MagiRogue/GameSys/Magic/Effects/DamageEffect.cs
@@ -13,8 +13,7 @@ namespace MagiRogue.GameSys.Magic.Effects
         public DamageTypes SpellDamageType { get; set; }
         public int BaseDamage { get; set; }
         public int Radius { get; set; }
-        public int ConeCircleSpan { get; set; }
-
+        public double ConeCircleSpan { get; set; }
         public bool TargetsTile { get; set; } = false;
         public EffectType EffectType { get; set; } = EffectType.DAMAGE;
         public bool IsHealing { get; set; }
